認証状態ごとのルーティング先の調整

+++ src/app/Http/Controllers/FavoriteController.php
@@ -11,11 +11,16 @@ class FavoriteController extends Controller
 {
     // お気に入り追加
     public function favoriteOn(Request $request){
-      // 新しくお気に入りアイテムを作成
-      Favorite::create([
-        'user_id' => Auth::user()->id,
-        'store_id' => $request->store_id,
-      ]);
+      if(Auth::user()) {
+        // 新しくお気に入りアイテムを作成
+        Favorite::create([
+          'user_id' => Auth::user()->id,
+          'store_id' => $request->store_id,
+        ]);
+      }
+      else {
+        session([Store::find($request->store_id)->name => true]);
+      }
 
       // 画面を更新
       if(strpos($_SERVER['HTTP_REFERER'],'mypage') === false){
@@ -28,9 +33,14 @@ class FavoriteController extends Controller
 
     // お気に入り削除
     public function favoriteOff(Request $request){
-      // 該当のお気に入りアイテムを削除
-      $favorite = Favorite::getFavorite(Auth::user()->id, $request->store_id);
-      $favorite->delete(); 
+      if(Auth::user()){
+        // 該当のお気に入りアイテムを削除
+        $favorite = Favorite::getFavorite(Auth::user()->id, $request->store_id);
+        $favorite->delete();
+      }
+      else {
+        session()->forget(Store::find($request->store_id)->name);
+      }
 
       // 画面を更新
       if(strpos($_SERVER['HTTP_REFERER'],'mypage') === false){
